CH Refactor: Rename *-data.ts to model.ts
