+++ src/blocks/content-helper/performance-details/model.ts
